review: move BC init to SamlConfiguration
