+++ server/src/test/java/org/cloudfoundry/identity/uaa/provider/saml/OpenSaml4AuthenticationProviderUnitTests.java
@@ -96,7 +96,7 @@ class OpenSaml4AuthenticationProviderUnitTests {
 
     @BeforeEach
     void setUp() {
-        IdentityZoneConfig.setupOpenSaml();
+        SamlConfiguration.setupOpenSaml();
     }
 
     @AfterEach void cleanup() {
