+++ uaa/src/test/java/org/cloudfoundry/identity/uaa/YamlConfigurationValidationTests.java
@@ -2,7 +2,7 @@ package org.cloudfoundry.identity.uaa;
 
 import org.cloudfoundry.identity.uaa.impl.config.YamlConfigurationValidator;
 import org.cloudfoundry.identity.uaa.impl.config.YamlServletProfileInitializer;
-import org.cloudfoundry.identity.uaa.provider.saml.IdentityZoneConfig;
+import org.cloudfoundry.identity.uaa.provider.saml.SamlConfiguration;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.BeansException;
@@ -49,7 +49,7 @@ class YamlConfigurationValidationTests {
     }
 
     private static TestApplicationContext createApplicationContext() {
-        IdentityZoneConfig.setupOpenSaml();
+        SamlConfiguration.setupOpenSaml();
         var applicationContext = new TestApplicationContext();
         var servletContext = new TestMockContext();
         applicationContext.setServletContext(servletContext);
